+++ backend/src/main/java/org/example/backend/professor/controller/ProfessorController.java
@@ -7,8 +7,8 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.example.backend.common.dto.PageRequestDto;
 import org.example.backend.common.dto.ResponseDto;
-import org.example.backend.professor.domain.dto.professor.ProfessorReqDto;
-import org.example.backend.professor.domain.dto.professor.ProfessorResDto;
+import org.example.backend.professor.domain.dto.ProfessorReqDto;
+import org.example.backend.professor.domain.dto.ProfessorResDto;
 import org.example.backend.professor.service.ProfessorService;
 import org.example.backend.thesis.domain.dto.ThesisResDto;
 import org.springframework.data.domain.Page;
