+++ backend/src/test/java/org/example/backend/professor/service/ProfessorServiceTest.java
@@ -1,8 +1,8 @@
 package org.example.backend.professor.service;
 
 import org.example.backend.IntegrationTestSupport;
-import org.example.backend.professor.domain.dto.professor.ProfessorReqDto;
-import org.example.backend.professor.domain.dto.professor.ProfessorResDto;
+import org.example.backend.professor.domain.dto.ProfessorReqDto;
+import org.example.backend.professor.domain.dto.ProfessorResDto;
 import org.example.backend.professor.domain.entity.Professor;
 import org.example.backend.professor.exception.ProfessorException;
 import org.example.backend.professor.exception.ProfessorExceptionType;
